:wrench: refactor (adjoint): Merge class and constructor files.
